+++ frontend/src/modules/app/constants.ts
